+++ docs/tutorials/classification/dive_deep_cifar10.py
@@ -240,10 +240,10 @@ def test(ctx, val_data):
 # Following is the script.
 #
 # .. note::
-#   In order to finish the tutorial quickly, we only train for 5 epochs.
+#   In order to finish the tutorial quickly, we only train for 3 epochs.
 #   In your experiments, we recommend setting ``epochs=240``.
 
-epochs = 5
+epochs = 3
 lr_decay_count = 0
 
 for epoch in range(epochs):
