+++ scripts/classification/imagenet/train_imagenet.py
@@ -1,9 +1,8 @@
-import argparse, time, logging
+import argparse, time, logging, os
 
 import mxnet as mx
 from mxnet import gluon, nd
 from mxnet import autograd as ag
-from utils import get_data_rec, get_data_loader
 from mxnet.gluon import nn
 from mxnet.gluon.data.vision import transforms
 
@@ -15,7 +14,7 @@ from gluoncv.utils import makedirs, LRScheduler
 parser = argparse.ArgumentParser(description='Train a model for image classification.')
 parser.add_argument('--data-dir', type=str, default='~/.mxnet/datasets/imagenet',
                     help='training and validation pictures to use.')
-parser.add_argument('--rec-train', type=str, default='~/.mxnet/datasets/imagenet/rec/train.rec', 
+parser.add_argument('--rec-train', type=str, default='~/.mxnet/datasets/imagenet/rec/train.rec',
                     help='the training data')
 parser.add_argument('--rec-train-idx', type=str, default='~/.mxnet/datasets/imagenet/rec/train.idx',
                     help='the index of training data')
@@ -118,6 +117,100 @@ if opt.dtype != 'float32':
 
 net = get_model(model_name, **kwargs)
 net.cast(opt.dtype)
+
+# Two functions for reading data from record file or raw images
+def get_data_rec(rec_train, rec_train_idx, rec_val, rec_val_idx, batch_size, num_workers):
+    rec_train = os.path.expanduser(rec_train)
+    rec_train_idx = os.path.expanduser(rec_train_idx)
+    rec_val = os.path.expanduser(rec_val)
+    rec_val_idx = os.path.expanduser(rec_val_idx)
+    jitter_param = 0.4
+    lighting_param = 0.1
+    mean_rgb = [123.68, 116.779, 103.939]
+    std_rgb = [58.393, 57.12, 57.375]
+
+    def batch_fn(batch, ctx):
+        data = gluon.utils.split_and_load(batch.data[0], ctx_list=ctx, batch_axis=0)
+        label = gluon.utils.split_and_load(batch.label[0], ctx_list=ctx, batch_axis=0)
+        return data, label
+
+    train_data = mx.io.ImageRecordIter(
+        path_imgrec         = rec_train,
+        path_imgidx         = rec_train_idx,
+        preprocess_threads  = num_workers,
+        shuffle             = True,
+        batch_size          = batch_size,
+
+        data_shape          = (3, 224, 224),
+        mean_r              = mean_rgb[0],
+        mean_g              = mean_rgb[1],
+        mean_b              = mean_rgb[2],
+        std_r               = std_rgb[0],
+        std_g               = std_rgb[1],
+        std_b               = std_rgb[2],
+        rand_mirror         = True,
+        random_resized_crop = True,
+        max_aspect_ratio    = 4. / 3.,
+        min_aspect_ratio    = 3. / 4.,
+        max_random_area     = 1,
+        min_random_area     = 0.08,
+        brightness          = jitter_param,
+        saturation          = jitter_param,
+        contrast            = jitter_param,
+        pca_noise           = lighting_param,
+    )
+    val_data = mx.io.ImageRecordIter(
+        path_imgrec         = rec_val,
+        path_imgidx         = rec_val_idx,
+        preprocess_threads  = num_workers,
+        shuffle             = False,
+        batch_size          = batch_size,
+
+        resize              = 256,
+        data_shape          = (3, 224, 224),
+        mean_r              = mean_rgb[0],
+        mean_g              = mean_rgb[1],
+        mean_b              = mean_rgb[2],
+        std_r               = std_rgb[0],
+        std_g               = std_rgb[1],
+        std_b               = std_rgb[2],
+    )
+    return train_data, val_data, batch_fn
+
+def get_data_loader(data_dir, batch_size, num_workers):
+    normalize = transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
+    jitter_param = 0.4
+    lighting_param = 0.1
+
+    def batch_fn(batch, ctx):
+        data = gluon.utils.split_and_load(batch[0], ctx_list=ctx, batch_axis=0)
+        label = gluon.utils.split_and_load(batch[1], ctx_list=ctx, batch_axis=0)
+        return data, label
+
+    transform_train = transforms.Compose([
+        transforms.RandomResizedCrop(224),
+        transforms.RandomFlipLeftRight(),
+        transforms.RandomColorJitter(brightness=jitter_param, contrast=jitter_param,
+                                     saturation=jitter_param),
+        transforms.RandomLighting(lighting_param),
+        transforms.ToTensor(),
+        normalize
+    ])
+    transform_test = transforms.Compose([
+        transforms.Resize(256),
+        transforms.CenterCrop(224),
+        transforms.ToTensor(),
+        normalize
+    ])
+
+    train_data = gluon.data.DataLoader(
+        imagenet.classification.ImageNet(data_dir, train=True).transform_first(transform_train),
+        batch_size=batch_size, shuffle=True, last_batch='discard', num_workers=num_workers)
+    val_data = gluon.data.DataLoader(
+        imagenet.classification.ImageNet(data_dir, train=False).transform_first(transform_test),
+        batch_size=batch_size, shuffle=False, num_workers=num_workers)
+
+    return train_data, val_data, batch_fn
 
 if opt.use_rec:
     train_data, val_data, batch_fn = get_data_rec(opt.rec_train, opt.rec_train_idx,
